org.jsonschema2pojo.ContentResolverTest require internet connection (#704)
