+++ jsonschema2pojo-core/src/test/java/org/jsonschema2pojo/ContentResolverTest.java
@@ -16,7 +16,8 @@
 
 package org.jsonschema2pojo;
 
-import static org.junit.Assert.*;
+import com.fasterxml.jackson.databind.JsonNode;
+import org.junit.Test;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -24,57 +25,18 @@ import java.io.IOException;
 import java.io.OutputStream;
 import java.net.URI;
 
-import org.apache.commons.io.IOUtils;
-import org.jsonschema2pojo.util.LocalHttpServerBuilder;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
-import org.junit.Test;
-
-import com.fasterxml.jackson.databind.JsonNode;
 import static org.hamcrest.Matchers.*;
+import static org.junit.Assert.assertThat;
 
 public class ContentResolverTest {
 
-    private ContentResolver resolver = new ContentResolver(); 
+    private ContentResolver resolver = new ContentResolver();
     
     @Test(expected=IllegalArgumentException.class)
     public void wrongProtocolCausesIllegalArgumentException() {
 
         URI uriWithUnrecognisedProtocol = URI.create("foobar://schema/address.json"); 
         resolver.resolve(uriWithUnrecognisedProtocol);
-    }
-
-    private static LocalHttpServerBuilder.Server server = null;
-
-    @BeforeClass
-    public static void beforeClass() throws Exception{
-        server = LocalHttpServerBuilder.createServer(
-                LocalHttpServerBuilder.context("/address", "application/json", "utf-8",
-                        IOUtils.toByteArray(ContentResolverTest.class.getResourceAsStream("/schema/address.json")))
-        );
-        server.startInRange(1024, 100);
-    }
-
-    @AfterClass
-    public static void  afterClass() throws Exception{
-        if(server!=null)
-            server.close();
-    }
-
-    @Test(expected=IllegalArgumentException.class)
-    public void brokenLinkCausesIllegalArgumentException() {
-
-        URI brokenHttpUri = URI.create("http://localhost:" + server.getPort() + "/sserdda");
-        resolver.resolve(brokenHttpUri);
-    }
-    
-    @Test
-    public void httpLinkIsResolvedToContent() {
-
-        URI httpUri = URI.create("http://localhost:" + server.getPort() + "/address");
-        JsonNode uriContent = resolver.resolve(httpUri);
-        
-        assertThat(uriContent.path("description").asText().length(), is(greaterThan(0)));
     }
 
     @Test
