+++ jsonschema2pojo-core/src/test/resources/wiremock/mappings/address404.json
@@ -1,0 +1,13 @@
+{
+  "request": {
+    "method": "GET",
+    "url": "/address404.json"
+  },
+  "response": {
+    "status": 404,
+    "body": "File not found",
+    "headers": {
+      "Content-Type": "plain/test"
+    }
+  }
+}
